requirements.txt
certifi==2020.6.20
chardet==3.0.4
dnspython==2.0.0
idna==2.10
requests==2.24.0
requests-file==1.5.1
six==1.15.0
termcolor==1.1.0
tldextract==2.2.3
urllib3==1.25.10
whois==0.9.7
